pre_analysis_py/requirements.txt
-e git+https://github.com/angr/ailment@638496df6547662517f5f524c9bde4b53fcbd176#egg=ailment
androguard==3.3.5
-e git+https://github.com/angr/angr@99b908dae67b789006384bd79fd14e379d8d68fb#egg=angr
-e git+https://github.com/angr/angr-management@b668b8649028ccb60c5da8dd6845a905cb61f10c#egg=angr_management
-e git+https://github.com/angr/angr-targets@31216c60992caad21a82e99be27676a2da9bbe08#egg=angr_targets
-e git+https://github.com/axt/angr-utils@7fd826a16b511727148f727150cea1d7ba6985f5#egg=angr_utils
-e git+https://github.com/angr/angrop@4a1664576db8b21eeb55660b3d8de8b041ef08d8#egg=angrop
-e git+https://github.com/angr/archinfo@bd84450115db29bef89022e3c63c7ea84facd46e#egg=archinfo
-e git+https://github.com/angr/archr@8c39c88ae834db3e88289ca3c9560c1b6063b01b#egg=archr
asn1crypto==1.4.0
astroid==2.9.3
asttokens==2.0.5
avatar2==1.4.6
backcall==0.2.0
-e git+https://github.com/axt/bingraphvis@0662ec6aaaa08c392b1f16d3fa8dffbdafe17466#egg=bingraphvis
bitstring==3.1.9
black==21.12b0
cachetools==5.0.0
capstone==4.0.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
-e git+https://github.com/angr/claripy@4c961b4dc664706be8142fe4868f27655bc8da77#egg=claripy
-e git+https://github.com/angr/cle@7dea3e72a06c7596b9ef7b884c42cb19bca7620a#egg=cle
click==8.0.3
colorama==0.4.4
configparser==5.2.0
coverage==6.3
CppHeaderParser==2.7.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
docker==5.0.3
dpkt==1.9.7.2
entrypoints==0.3
executing==0.8.2
flaky==3.7.0
fonttools==4.29.1
future==0.18.2
getmac==0.8.3
gitdb==4.0.9
GitPython==3.1.26
greenlet==1.1.2
idna==3.3
intervaltree==3.1.0
ipdb==0.13.9
ipykernel==6.7.0
ipython==8.0.1
ipython-genutils==0.2.0
isort==5.10.1
itanium-demangler==1.0
jedi==0.18.1
jupyter-client==7.1.2
jupyter-core==4.9.1
keystone-engine==0.9.2
kiwisolver==1.3.2
lazy-object-proxy==1.7.1
lxml==4.8.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mpmath==1.2.1
mulpyplexer==0.9
mypy-extensions==0.4.3
nampa==0.1.1
nclib==1.0.1
nest-asyncio==1.5.4
networkx==2.6.3
ninja==1.10.2.3
nose==1.3.7
nose-timer==1.0.1
npyscreen==4.10.5
numpy==1.22.2
packaging==21.3
parse==1.19.0
parso==0.8.3
patchelf-wrapper==1.2.0
pathspec==0.9.0
pefile==2021.9.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pkg_resources==0.0.0
platformdirs==2.4.1
plumbum==1.7.2
ply==3.11
posix-ipc==1.0.5
progressbar==2.5
progressbar2==4.0.0
prompt-toolkit==3.0.26
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydot==1.4.2
pyelftools==0.27
pygdbmi==0.10.0.1
Pygments==2.11.2
pygraphviz==1.9
pylink-square==0.11.1
pylint==2.12.2
pyparsing==3.0.7
pyqodeng.core==0.0.4
PySide2==5.15.2.1
PySMT==0.9.1.dev139
PySocks==1.7.1
-e git+https://github.com/angr/pysoot@d08dc569ec35796ccea5509b3e04b74967bcfd48#egg=pysoot
python-dateutil==2.8.2
python-utils==3.1.0
pyusb==1.2.1
-e git+https://github.com/angr/pyvex@13f407498daa83882abe82b1598482a0ba69be6d#egg=pyvex
pyxdg==0.27
pyzmq==22.3.0
QtAwesome==1.1.1
qtconsole==5.2.2
QtPy==2.0.0
qtterm==0.5.1
requests==2.27.1
rpyc==5.0.1
shellphish-qemu==0.12.3
shiboken2==5.15.2.1
six==1.16.0
smmap==5.0.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.31
stack-data==0.1.4
sympy==1.9
toml==0.10.2
tomli==1.2.3
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.0.1
unicorn==1.0.2rc4
urllib3==1.26.8
wcwidth==0.2.5
websocket-client==1.2.3
wrapt==1.13.3
z3-solver==4.8.14.0
